+++ projects/owa-data/owa/data/datasets/transforms/tokenized.py
@@ -1,0 +1,6 @@
+"""Tokenized transform for OWA datasets."""
+
+
+def create_tokenized_transform():
+    """Create transform for TOKENIZED stage."""
+    return lambda batch: batch
